requirements.txt
discord.py==2.4.0
openai==1.12.0
google-generativeai==0.8.3
python-dotenv==1.0.1
aiohttp==3.9.5
sqlalchemy==2.0.28
aiosqlite==0.19.0
